refactor: 디렉토리 구조 변경 및 final 누락 변수 final 추가
