+++ src/main/java/domain/lotto/LottoGame.java
@@ -1,10 +1,10 @@
 package domain.lotto;
 
+import domain.lotto.number.WinningNumbers;
+import domain.lotto.result.LottoChecker;
 import domain.lotto.result.LottoResults;
 import domain.lotto.ticket.LottoTicket;
 import domain.lotto.ticket.LottoTickets;
-
-import java.util.List;
 
 public class LottoGame {
     private final LottoTickets lottoTickets;
