+++ src/test/java/LottoCheckerTest.java
@@ -1,5 +1,5 @@
-import domain.lotto.LottoChecker;
-import domain.lotto.WinningNumbers;
+import domain.lotto.result.LottoChecker;
+import domain.lotto.number.WinningNumbers;
 import domain.lotto.result.LottoResultType;
 import domain.lotto.ticket.LottoTicket;
 import org.junit.jupiter.api.DisplayName;
